Refactored board to be a matrix of tetramino.
Refactoring done to prepare for printing rotated tetramino.  Refactored the
board to be a matrix of tetramino.  This included some string-to-tetramino
cell mappings et al.  Rotation was added as a record field, but is not
implemented yet.  Some literals and magic numbers were removed, and more
library stuff was used.